Add migrations and entities to separate files for accounting, fix ack.
